% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.phylo.hclust.node.attributes.R
\name{as.phylo.hclust.node.attributes}
\alias{as.phylo.hclust.node.attributes}
\title{as.phylo.hclust.node.attributes}
\usage{
\method{as.phylo}{hclust.node.attributes}(x, attribute)
}
\arguments{
\item{x}{hclust object}

\item{attribute}{attributes for nodes}
}
\description{
Function to combine smaller clusters from a fast hierarchical algorithm to maximise the BAPS likelihood.
}
