% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_baps_partition.R
\name{best_baps_partition}
\alias{best_baps_partition}
\title{best_baps_partition}
\usage{
best_baps_partition(sparse.data, phylo)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}

\item{phylo}{a phylo object representing the hierarchical clustering that is to be cut}
}
\value{
a final clustering
}
\description{
Function to combine smaller clusters from a fast hierarchical algorithm to maximise the BAPS likelihood.
}
\examples{

fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
sim.matrix <- snp_similarity(sparse.data)
x <- as.dist(1-sim.matrix/max(sim.matrix))
phylo <- ape::as.phylo(hclust(x, method="average"))
partition <- best_baps_partition(sparse.data, phylo)

phylo <- phytools::read.newick(system.file("extdata", "seqs.fa.treefile", package = "fastbaps"))
partition <- best_baps_partition(sparse.data, phylo)

}
