% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_fast_baps.R
\name{boot_fast_baps}
\alias{boot_fast_baps}
\title{boot_fast_baps}
\usage{
boot_fast_baps(sparse.data, k.init = NULL, n.replicates = 100,
  hc.method = "ward", n.cores = 1, quiet = TRUE)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}

\item{k.init}{the initial number of clusters to start the bayesian hierarchical clustering from. Defaults to (number of sequences)/4}

\item{n.replicates}{the number of bootstrap replicates to perform (default=100)}

\item{hc.method}{the type of initial hierarchical clustering to use. Can be with 'ward' or 'genie' (default='ward')}

\item{n.cores}{the number of cores to use in clustering}

\item{quiet}{whether or not to print progress information (default=FALSE)}
}
\value{
a co-occurence count matrix
}
\description{
Function to perform bootstrap replicated of fastbaps.
}
\examples{

fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
boot.result <- boot_fast_baps(sparse.data)
dendro <- as.dendrogram(fast_baps(sparse.data))
heatmap(boot.result, dendro, dendro)


}
