% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_marginal_llk.R
\name{calc_marginal_llk}
\alias{calc_marginal_llk}
\title{calc_marginal_llk}
\usage{
calc_marginal_llk(sparse.data, partition)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}

\item{partition}{the partition for which the log marginal likelihood is to be calculated}
}
\value{
the log marginal likelihood
}
\description{
Function to calculate the log marginal likelihood of a given partition
}
\examples{
fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
partition <- 1:ncol(sparse.data$snp.matrix)
calc_marginal_llk(sparse.data, partition)

}
