% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_baps.R
\name{fast_baps}
\alias{fast_baps}
\title{fast_baps}
\usage{
fast_baps(
  sparse.data,
  k.init = NULL,
  hc.method = "ward",
  n.cores = 1,
  quiet = FALSE
)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}

\item{k.init}{the initial number of clusters to start the bayesian hierarchical clustering from. Defaults to (number of sequences)/4}

\item{hc.method}{the type of initial hierarchical clustering to use. Can be with 'ward' or 'genie' (default='ward')}

\item{n.cores}{the number of cores to use in clustering}

\item{quiet}{whether or not to print progress information (default=FALSE)}
}
\value{
a final hierarchical clustering clustering as a hclust object.
}
\description{
Function to perform Bayesian hierarchical clustering of population structure.
}
\examples{

fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
system.time({baps.hc <- fast_baps(sparse.data)})

}
