% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_cluster_subset.R
\name{find_max_cluster_subset}
\alias{find_max_cluster_subset}
\title{find_max_cluster_subset}
\usage{
find_max_cluster_subset(phylo, node.weights)
}
\description{
Function to find the best partition given a weighting on the nodes
}
