% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_clusters.R
\name{fix_clusters}
\alias{fix_clusters}
\title{fix_clusters}
\usage{
fix_clusters(sparse.data, clusters, n.iterations = 1)
}
\arguments{
\item{n.iterations}{}
}
\value{
a final clustering
}
\description{
Function to iteratively update an intially clustering, greedily maximising the BAPS marginal likelihood.
}
\examples{

fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
sim.matrix <- snp_similarity(t(sparse.data$snp.matrix>0))
x <- as.dist(1-sim.matrix/max(sim.matrix))
phylo <- ape::as.phylo(hclust(x, method="average"))
clusters <- best_baps_partition_fllk(sparse.data, phylo)
clusters <- fix_clusters(sparse.data, clusters)

}
