% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hclust.R
\name{get_hclust}
\alias{get_hclust}
\title{as.phylo.hclust.node.attributes
get_hclust}
\usage{
get_hclust(sparse.data, quiet, method = "ward", n.cores = 1)
}
\arguments{
\item{sparse.data}{sparse data object generated using}

\item{quiet}{whether to supress additional printing information}

\item{method}{which hierarchical clustering method. One of 'ward' or 'genie'. (default='ward')}

\item{n.cores}{n.cores to use (not implemented)}
}
\description{
Function to generate hclust object. Uses initial kmeans if dataset is too large.
}
