% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iorder.R
\name{iorder}
\alias{iorder}
\title{iorder}
\usage{
iorder(m)
}
\arguments{
\item{m}{}
}
\value{
an ordering
}
\description{
A function used to order a hclust object for plotting. Taken directly from https://github.com/bwlewis/hclust_in_R/blob/master/hc.R
}
