% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_level_best_baps_partition.R
\name{multi_level_best_baps_partition}
\alias{multi_level_best_baps_partition}
\title{multi_level_best_baps_partition}
\usage{
multi_level_best_baps_partition(
  sparse.data,
  h,
  levels = 2,
  n.cores = 1,
  quiet = TRUE
)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}

\item{h}{a hclust object representing the hierarchical clustering that is to be cut}

\item{levels}{the number of levels to investigate (default=2)}

\item{n.cores}{the number of cores to use in clustering}

\item{quiet}{whether to print additional information or not (default=TRUE)}
}
\value{
a data.frame representing the final clustering at multiple resolutions
}
\description{
Function to partition a given phylogeny, choosing the best partition at each level using the fastbaps approach.
}
\examples{

fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
d <- snp_dist(sparse.data)
d <- as.dist(d/max(d))
h <- hclust(d, method="ward.D2")
multi.res.df <- multi_level_best_baps_partition(sparse.data, h, levels=2)

}
