% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_baps.R
\name{new_baps}
\alias{new_baps}
\title{hc_baps}
\usage{
new_baps(sparse.data, ncores = 1, k.init = NULL)
}
\arguments{
\item{snp.matrix}{}
}
\value{
a final clustering
}
\description{
Function to perform Bayesian hierarchical clustering of population structure.
}
\examples{

fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
k.init = 100
gp.hc <- new_baps(sparse.data, k.init=k.init)
best.partition <- best_baps_partition_fllk(sparse.data, phylo = ape::as.phylo(gp.hc))

}
