% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_prior.R
\name{optimise_prior}
\alias{optimise_prior}
\title{optimise_prior}
\usage{
optimise_prior(
  sparse.data,
  grid.interval = c(5e-04, 10),
  type = "optimise.baps",
  hc.method = "ward",
  n.cores = 1
)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}

\item{grid.interval}{the upper and lower bound for the hyperparameter to optimise over (default=c(5e-4, 10))}

\item{type}{one of 'optimise.symmetric','symmetric', 'optimise.baps', 'baps' or 'hc'. (default='optimise.baps')}

\item{hc.method}{the type of initial hierarchical clustering to use. Can be with 'ward' or 'genie' (default='ward')}

\item{n.cores}{number of cores to use (currently not implemented)}
}
\value{
a sparse.data object with the prior optimised via grid search
}
\description{
Function to optimise the prior using Bayes Factors and grid search
}
\examples{
fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
sparse.data <- optimise_prior(sparse.data)

}
