% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_dist.R
\name{snp_dist}
\alias{snp_dist}
\title{snp_dist}
\usage{
snp_dist(sparse.data)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}
}
\value{
A pairwise snp distance matrix
}
\description{
Function to calculate pairwise snp distance matrix
}
\examples{
fasta.file <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
snp_dist(sparse.data)

}
