% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_similarity.R
\name{snp_similarity}
\alias{snp_similarity}
\title{snp_similarity}
\usage{
snp_similarity(sparse.data)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}
}
\value{
A pairwise snp similarity matrix
}
\description{
Function to calculate pairwise snp similarity matrix
}
\examples{
fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
snp_similarity(sparse.data)

}
