% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_baps_partition.R
\name{best_baps_partition}
\alias{best_baps_partition}
\title{best_baps_partition}
\usage{
best_baps_partition(sparse.data, h, quiet = FALSE)
}
\arguments{
\item{sparse.data}{a sparse SNP data object returned from import_fasta_sparse_nt}

\item{h}{a hclust object representing the hierarchical clustering that is to be cut}

\item{quiet}{suppress the printing of extra information (default=FALSE)}
}
\value{
a final clustering
}
\description{
Function to combine smaller clusters from a fast hierarchical algorithm to maximise the BAPS likelihood.
}
\examples{

fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)
d <- snp_dist(sparse.data)
d <- as.dist(d/max(d))
h <- hclust(d, method="ward.D2")
partition <- best_baps_partition(sparse.data, h)

newick.file.name <- system.file("extdata", "seqs.fa.treefile", package = "fastbaps")
iqtree <- phytools::read.newick(newick.file.name)
h <- phytools::midpoint.root(iqtree)
best.partition <- best_baps_partition(sparse.data, h)

}
