% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fasta_sparse_nt.R
\name{import_fasta_sparse_nt}
\alias{import_fasta_sparse_nt}
\title{import_fasta_sparse_nt}
\usage{
import_fasta_sparse_nt(fasta, prior = "baps", check.fasta = TRUE)
}
\arguments{
\item{fasta}{path to the fasta file or an ape DNAbin object}

\item{prior}{the type of prior to use. Can be one of 'baps' and 'mean' (default=baps)}

\item{check.fasta}{whether to check the fasta file for issue. Slows things down a little but may be avoided for very large fasta files.}
}
\value{
A sparse matrix reprsentation of the SNPs (different to the consensus sequence)
}
\description{
Imports a fasta file to a sparse matrix representing SNPs from the consensus
}
\examples{
fasta <- system.file("extdata", "seqs.fa", package = "fastbaps")
fasta <- ape::read.FASTA(fasta)
sparse.data <- import_fasta_sparse_nt(fasta)

}
