% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fasta_sparse_nt.R
\name{import_fasta_sparse_nt}
\alias{import_fasta_sparse_nt}
\title{import_fasta_sparse_nt}
\usage{
import_fasta_sparse_nt(fasta.file.name, prior = "baps")
}
\arguments{
\item{fasta.file.name}{path to the fasta file}

\item{prior}{the type of prior to use. Can be one of 'baps' and 'mean' (default=baps)}
}
\value{
A sparse matrix reprsentation of the SNPs (different to the consensus sequence)
}
\description{
Imports a fasta file to a sparse matrix representing SNPs from the consensus
}
\examples{
fasta.file.name <- system.file("extdata", "seqs.fa", package = "fastbaps")
sparse.data <- import_fasta_sparse_nt(fasta.file.name)

}
